Add BaseActivity and ArtistsFeedActivity
