+++ MyTattooArtist/.idea/misc.xml
@@ -5,7 +5,9 @@
       <map>
         <entry key="..\:/Users/Joonas/Koodit/mobile-project/MyTattooArtist/app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="..\:/Users/Joonas/Koodit/mobile-project/MyTattooArtist/app/src/main/res/layout/artist_profile.xml" value="0.1" />
+        <entry key="..\:/Users/Joonas/Koodit/mobile-project/MyTattooArtist/app/src/main/res/layout/artists_feed.xml" value="0.14722222222222223" />
         <entry key="..\:/Users/Joonas/Koodit/mobile-project/MyTattooArtist/app/src/main/res/layout/sign_up.xml" value="0.1" />
+        <entry key="..\:/Users/Joonas/Koodit/mobile-project/MyTattooArtist/app/src/main/res/layout/user_profile.xml" value="0.14722222222222223" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.15052083333333333" />
         <entry key="app/src/main/res/menu/menu.xml" value="0.16458333333333333" />
       </map>
